Add replacing NewVaccineModulePage -> VaccinePage
